Move filesystem and store unit tests to src/tests/
filesystem.rs and store.rs were both getting unwieldy. Extract their
in-crate test modules to dedicated files under src/tests/ wired via
#[cfg(test)] #[path] so they remain children of their parent modules
and keep crate-private access. Leave gql tests inline — they're small
and live inside nested inline modules where #[path] toward src/tests/
would need to traverse virtual directories that do not resolve.
